general | type adjustments + test fixes (currently failing - sync).

+++ include/liblinal/liblinal.h
@@ -50,13 +50,13 @@ typedef struct {
 
 bool la_matrix_is_equal(LA_Matrix *m1, LA_Matrix *m2, bool *result);
 
-bool la_matrix_get(LA_Matrix *m, int row, int column, float *result);
+bool la_matrix_get(LA_Matrix *m, size_t row, size_t column, float *result);
 
-bool la_matrix_set(LA_Matrix *m, int row, int column, float value);
+bool la_matrix_set(LA_Matrix *m, size_t row, size_t column, float value);
 
-bool la_matrix_get_row(LA_Matrix *m, int row, LA_Matrix *result);
+bool la_matrix_get_row(LA_Matrix *m, size_t row, LA_Matrix *result);
 
-bool la_matrix_get_column(LA_Matrix *m, int column, LA_Matrix *result);
+bool la_matrix_get_column(LA_Matrix *m, size_t column, LA_Matrix *result);
 
 // matrix arithmetic.
